Finally got some real stack traces going
Only works with libbacktrace which is... uh created by the dude that did
go generics and then promptly left google. So not sure how maintained
its going to be.  Actually the whole story with symbolizers and stack
traces is kinda hilarious. For now this works on linux and maybe mac,
but definitely not windows.  will need to come back and amend it in the
future