Fix for connection pool leak
Connection's werent properly reset when disconnecting leading to a
memory leak, so with this fix the server is able to handle much greater
loads than before.

Now ofcourse this an inflection point, as its good enough at single
threaded stuff, so might be good next step to boost the performance a
bit by spinning up multiple event loops. But that can be saved as a
later exercise.